[deckhouse] fix
